Updated files so everything works as it should.

+++ app/src/main/java/hr/foi/techtitans/ttpay/products/activity_products/UpdateServiceActivity.kt
@@ -159,7 +159,7 @@ class UpdateServiceActivity : AppCompatActivity() {
             updatedFields["description"] = descriptionEditText.text.toString()
         }
         if (isFieldChanged(priceEditText, originalPrice)) {
-            updatedFields["price"] = priceEditText.text.toString().toInt()
+            updatedFields["price"] = priceEditText.text.toString().toDouble()
         }
         if (isFieldChanged(currencyEditText, originalCurrency)) {
             updatedFields["currency"] = currencyEditText.text.toString()
